+++ src/design/plone/contenttypes/tests/test_ct_servizio.py
@@ -74,6 +74,7 @@ class TestServizio(unittest.TestCase):
                 "collective.dexteritytextindexer",
                 "plone.translatable",
                 "kitconcept.seo",
+                "plone.versioning",
             ),
         )