+++ src/design/plone/contenttypes/upgrades/upgrades.py
@@ -891,6 +891,43 @@ def to_5220(context):
 
 
 def to_5300(context):
+    update_profile(context, "plone-difftool")
+    update_profile(context, "repositorytool")
 
+    portal_types = api.portal.get_tool(name="portal_types")
+    for portal_type, fti in portal_types.items():
+        if portal_type in [
+            "CartellaModulistica",
+            "Documento",
+            "Link",
+            "Pagina Argomento",
+            "Persona",
+            "Servizio",
+            "UnitaOrganizzativa",
+            "Venue",
+        ]:
+            behaviors = list(getattr(fti, "behaviors", ()))
+            if "plone.versioning" not in behaviors:
+                behaviors.append("plone.versioning")
+                fti.behaviors = tuple(behaviors)
+
+
+def to_5310(context):
+    """
+    Reindex Bandi
+    """
+    logger.info("Reindex SearchableText for Bandi items.")
+    catalog = api.portal.get_tool("portal_catalog")
+    i = 0
+    brains = catalog(portal_type="Bando")
+    tot = len(brains)
+    for brain in brains:
+        i += 1
+        if i % 500 == 0:
+            logger.info("Progress: {}/{}".format(i, tot))
+        brain.getObject().reindexObject(idxs=["SearchableText"])
+
+
+def to_5400(context):
     logger.info('Remove "volto.blocks" behavior from News Item and Event.')
     remove_blocks_behavior(context)